Move test utils out of "scripts" folder
