+++ .babelrc
@@ -36,7 +36,7 @@
           "pages": "./pages",
           "scripts": "./scripts",
           "static": "./static",
-          "test-utils": "./scripts/test-utils"
+          "test-utils": "./test-utils"
         }
       }
     ]
