+++ jest.config.js
@@ -78,7 +78,7 @@ module.exports = {
 
   // A map from regular expressions to module names that allow to stub out resources with a single module
   moduleNameMapper: {
-    '\\.(jpg|jpeg|png|gif|svg)$': '<rootDir>/scripts/test-utils/mocks/testFileMock.js',
+    '\\.(jpg|jpeg|png|gif|svg)$': '<rootDir>/test-utils/mocks/testFileMock.js',
     '\\.css$': 'identity-obj-proxy',
   },
